split indicore library
