+++ libs/indicore/base64.c
