+++ libs/indicore/base64_luts.h
