+++ libs/indicore/indicom.c
