+++ libs/indicore/indicom.h
