+++ libs/indicore/indililxml.h
