+++ libs/indicore/indiuserio.c
