+++ libs/indicore/libastro.cpp
