+++ libs/indicore/lilxml.cpp
