+++ libs/indicore/lilxml.h
