+++ libs/indicore/locale_compat.h
