+++ libs/indicore/sharedblob.h
