+++ libs/indicore/sharedblob_parse.cpp
