+++ libs/indicore/shm_open_anon.c
