+++ libs/indicore/userio.c
