+++ libs/indicore/userio.h
